PyCharmMiscProject/hla_project/requirements.txt
biopython>=1.83
requests>=2.31
httpx>=0.27
psycopg2-binary>=2.9
sqlalchemy>=2.0
openai>=1.30
pydantic>=2.5
neo4j>=5.17
pymupdf>=1.24
pdfplumber>=0.11
python-dotenv>=1.0
tenacity>=8.2
tqdm>=4.66
click>=8.1
